refactor: 문서 목록 반환 시 각 문서의 최신버전 포함하여 반환

+++ src/main/java/com/jobnote/domain/document/domain/DocumentVersion.java
@@ -18,7 +18,7 @@ public class DocumentVersion extends BaseTimeEntity {
     @Column(name = "document_version_id")
     private Long id;
 
-    private int version;
+    private Long version;
 
     private String title;
 
@@ -34,7 +34,7 @@ public class DocumentVersion extends BaseTimeEntity {
 
     @Builder
     public DocumentVersion(
-            final int version,
+            final Long version,
             final String title,
             final String originFileName,
             final String fileKey,
